cargo check no errors on api crate
